Seperated SaveAndLoadPerformer and CharacterSerializer
