+++ MazesAndMinotaurs/app/src/main/java/com/example/cis/mazeminotaurs/character/SaveAndLoadPerformer.java
@@ -4,14 +4,11 @@ import com.example.cis.mazeminotaurs.Armor;
 import com.example.cis.mazeminotaurs.Portfolio;
 import com.example.cis.mazeminotaurs.Weapon;
 import com.example.cis.mazeminotaurs.character.classes.Barbarian;
-import com.example.cis.mazeminotaurs.character.classes.BaseClass;
 import com.example.cis.mazeminotaurs.character.stats.Score;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
 import com.google.gson.reflect.TypeToken;
 
 import java.lang.reflect.Type;
@@ -83,40 +80,11 @@ public class SaveAndLoadPerformer {
     }
 
     public static void loadPortfolio(String jsonString) {
-        Gson gson = new Gson();
+        GsonBuilder builder = new GsonBuilder();
         Type listType = new TypeToken<ArrayList<PlayerCharacter>>() {
         }.getType();
-        ArrayList<PlayerCharacter> newPort = gson.fromJson(jsonString, listType);
+        builder.registerTypeAdapter(PlayerCharacter.class, new CharacterSerializer());
+        ArrayList<PlayerCharacter> newPort = builder.create().fromJson(jsonString, listType);
         Portfolio.get().setPortfolio(newPort);
-    }
-}
-
-/**
- * Custom serializer for transforming the PlayerCharacter class into json.
- */
-class CharacterSerializer implements JsonSerializer<PlayerCharacter> {
-    @Override
-    public JsonElement serialize(PlayerCharacter src, Type typeOfSrc, JsonSerializationContext context) {
-        BaseClass characterClass = src.getCharClass();
-        characterClass.setCharacter(null);
-
-        JsonObject rootObject = new JsonObject();
-        rootObject.add("mScores", context.serialize(src.getScores()));
-        rootObject.add("mCharClass", context.serialize(src.getCharClass()));
-        rootObject.add("mGender", context.serialize(src.getGender()));
-        rootObject.add("mMoney", context.serialize(src.getMoney()));
-        rootObject.add("mAge", context.serialize(src.getAge()));
-        rootObject.add("mName", context.serialize(src.getName()));
-        rootObject.add("mInventory", context.serialize(src.getInventory()));
-        rootObject.add("mCurrentWeapon", context.serialize(src.getCurrentWeapon()));
-        rootObject.add("mHelmet", context.serialize(src.getHelmet()));
-        rootObject.add("mBreastplate", context.serialize(src.getBreastplate()));
-        rootObject.add("mShield", context.serialize(src.getShield()));
-
-        JsonElement classJson = context.serialize(characterClass);
-        rootObject.add("mCharClass", classJson);
-
-        characterClass.setCharacter(src);
-        return rootObject;
     }
 }
